backend/requirements.txt
# Core web
Flask==3.1.2
Flask-SocketIO==5.4.1
flask-cors==5.0.0
python-dotenv==1.1.1

# Cloud / Firebase
firebase_admin==7.1.0

# HTTP / Async client (for Gemini or external APIs)
httpx==0.28.1

# Dev / Test (最小導入)
ruff==0.6.9
mypy==1.11.2
pytest==8.3.3
pytest-cov==5.0.0

# (legacy) 追加ライブラリは旧 `flask/requirements.txt` を参照しつつ精査予定
